beta version 0.2
Plenty of bugfixes for Linux kernel later than 3.5 and Ubuntu later than 10.10.
More organized code to improve portability.
Regression tests for Pal to test completeness of implementation.

+++ LibOS/shim/src/sys/shim_sandbox.c
@@ -300,7 +300,7 @@ struct cfg_arg {
     int offset;
 };
 
-static size_t __write (void * f, void * buf, size_t len)
+static int __write (void * f, void * buf, int len)
 {
     struct cfg_arg * arg = f;
